workspaces/MPWR_Update_Cancel_Agent/requirements.txt
playwright
requests
python-dotenv
pytz
supabase==2.28.3
